Correct roundFractional.
Merge _radix and _rational.
